pyproject.toml
[tool.poetry]
name = "class-excel"
version = "0.1.0"
description = ""
authors = ["rafi mahrus"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openpyxl = "^3.0.10"
pillow = "^11.1.0"
starlette = "^0.45.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
